requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
black==23.11.0
blinker==1.4
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
exceptiongroup==1.2.2
filelock==3.16.1
flake8==7.0.0
fonttools==4.55.4
fsspec==2024.12.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
grpcio==1.69.0
gunicorn==23.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.8.2
mccabe==0.7.0
more-itertools==8.10.0
mpmath==1.3.0
mypy==1.7.0
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
packaging==24.2
pandas==2.1.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==4.23.4
psycopg2==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.11.1
pycurl==7.44.1
pyflakes==3.2.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==7.4.4
pytest-order==1.2.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.15.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sqlparse==0.5.3
sympy==1.13.3
systemd-python==234
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.4.1+cu121
torchvision==0.19.1
tqdm==4.66.1
triton==3.0.0
Twisted==22.1.0
types-requests==2.31.0.10
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.3.0
wadllib==1.3.6
Werkzeug==3.1.3
zipp==1.0.0
zope.interface==5.4.0
